cria o crud da tabela tasks

+++ src/controllers/teams-controller.ts
@@ -31,7 +31,7 @@ class TeamsController {
     async create(request: Request, response: Response) {
         const bodySchema = z.object({
             name: z.string(),
-            description: z.string().optional()
+            description: z.string()
         })
         const { name, description } = bodySchema.parse(request.body)
         const teams = await prisma.teams.create({
@@ -56,6 +56,13 @@ class TeamsController {
         try {
             const { id } = paramsSchema.parse(request.params)
             const data = bodySchema.parse(request.body)
+            const findTask = await prisma.teams.findUnique({
+                where: { id }
+            })
+
+            if(!findTask) {
+                return response.status(404).json({ message: "ID não encontrado."})
+            }
 
             const teams = await prisma.teams.update({
                 where: { id },
@@ -63,24 +70,33 @@ class TeamsController {
             })
             return response.status(200).json({ message: "Registro atualizado com sucesso.", teams})
         } catch (error) {
-            return response.status(404).json({ message: "ID não encontrado."})
+            return response.status(400).json({ message: "Dados inválidos."})
         }
     }
 
     async delete(request: Request, response: Response) {
-        
+
         const paramsSchema = z.object({
             id: z.string().uuid()
         })
 
         try {
             const { id } = paramsSchema.parse(request.params)
+
+            const findTask = await prisma.teams.findUnique({
+                where: { id }
+            })
+
+            if(!findTask) {
+                return response.status(404).json({ message: "ID não encontrado"})
+            }
+
             await prisma.teams.delete({
                 where: { id }
             })
             return response.status(200).json({ message: "Registro removido com sucesso."}) 
         } catch (error) {
-            return response.status(404).json({ message: "ID não encontrado"})
+            return response.status(400).json({ message: "Dados inválidos."})
         }
     }
 }
